+++ src/log.h
@@ -37,7 +37,7 @@ typedef void (*logger_print_func)(enum loglevel level, const char*, va_list);
 void logger(enum loglevel level, const char *fmt, ...) __attribute__ ((format (printf, 2, 3)));
 int logger_set_logfile(const char* path);
 void logger_set_print_func(logger_print_func func);
-void logger_dump_hex(enum loglevel level, const void* buf, unsigned int len);
+void logger_dump_hex(enum loglevel level, const void* buf, size_t len);
 void logger_dump_plist(enum loglevel level, plist_t plist, int human_readable);
 
 #endif
